+++ frontend/src/lib/derived/sns-topics.derived.ts
@@ -7,12 +7,16 @@ import type {
   TopicInfoWithUnknown,
 } from "$lib/types/sns-aggregator";
 import { convertDtoToListTopicsResponse } from "$lib/utils/sns-aggregator-converters.utils";
-import { getSnsTopicsByProject } from "$lib/utils/sns-topics.utils";
 import type { Principal } from "@dfinity/principal";
-import { isNullish, nonNullish } from "@dfinity/utils";
+import {
+  fromDefinedNullable,
+  fromNullable,
+  isNullish,
+  nonNullish,
+} from "@dfinity/utils";
 import { derived, type Readable } from "svelte/store";
 
-export interface SnsTopicsStore {
+export interface SnsParametersStore {
   [rootCanisterId: RootCanisterIdText]:
     | ListTopicsResponseWithUnknown
     | undefined;
@@ -22,23 +26,45 @@ export interface SnsTopicsStore {
  * A store containing SNS topics for each project.
  * Returns undefined for projects where topics are not supported.
  */
-export const snsTopicsStore: Readable<SnsTopicsStore> = snsAggregatorDerived(
-  (sns) =>
+export const snsTopicsStore: Readable<SnsParametersStore> =
+  snsAggregatorDerived((sns) =>
     isNullish(sns?.topics)
       ? undefined
       : convertDtoToListTopicsResponse(sns.topics)
-);
+  );
 
+// TODO(sns-topics): Consider this to be a utility function.
 export const createSnsTopicsProjectStore = (
   rootCanisterId: Principal | null | undefined
 ): Readable<Array<TopicInfoWithUnknown> | undefined> =>
   derived<typeof snsTopicsStore, Array<TopicInfoWithUnknown> | undefined>(
     snsTopicsStore,
-    ($snsTopicStore) =>
-      getSnsTopicsByProject({
-        rootCanisterId,
-        snsTopicsStore: $snsTopicStore,
-      })
+    ($snsTopicStore) => {
+      const rootCanisterIdText = rootCanisterId?.toText();
+      if (isNullish(rootCanisterIdText)) {
+        return undefined;
+      }
+
+      const topicResponse = $snsTopicStore[rootCanisterIdText];
+      if (isNullish(topicResponse)) return undefined;
+
+      const topics = fromNullable(topicResponse.topics);
+      if (isNullish(topics)) return undefined;
+
+      // sorts topics with critical topics first, then alphabetically within each group
+      return topics.sort((a, b) => {
+        const isACritical = fromDefinedNullable(a.is_critical);
+        const isBCritical = fromDefinedNullable(b.is_critical);
+
+        if (isACritical && !isBCritical) return -1;
+        if (!isACritical && isBCritical) return 1;
+
+        const nameOfA = fromDefinedNullable(a.name);
+        const nameOfB = fromDefinedNullable(b.name);
+
+        return nameOfA.localeCompare(nameOfB);
+      });
+    }
   );
 
 export const createEnableFilteringBySnsTopicsStore = (
