+++ frontend/src/lib/pages/Portfolio.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-  import AdoptedProposalCard from "$lib/components/portfolio/AdoptedProposalCard.svelte";
   import HeldTokensCard from "$lib/components/portfolio/HeldTokensCard.svelte";
   import LaunchProjectCard from "$lib/components/portfolio/LaunchProjectCard.svelte";
   import LoginCard from "$lib/components/portfolio/LoginCard.svelte";
@@ -33,7 +32,6 @@
     tableProjects: TableProject[];
     snsProjects: SnsFullProject[];
     openSnsProposals: ProposalInfo[];
-    adoptedSnsProposals: SnsFullProject[];
   };
 
   const {
@@ -41,7 +39,6 @@
     tableProjects,
     snsProjects,
     openSnsProposals,
-    adoptedSnsProposals,
   }: Props = $props();
 
   const totalTokensBalanceInUsd = $derived(getTotalBalanceInUsd(userTokens));
@@ -94,6 +91,8 @@
   const areStakedTokensLoading = $derived(
     tableProjects.some((project) => project.isStakeLoading)
   );
+  // Determines the display state of the staked tokens card
+  // Similar logic to heldTokensCard but for staked tokens
   const stakedTokensCard: TokensCardType = $derived(
     !$authSignedInStore
       ? "full"
@@ -153,23 +152,7 @@
       }))
   );
 
-  const adoptedSnsProposalsCards = $derived(
-    [...adoptedSnsProposals]
-      .sort(comparesByDecentralizationSaleOpenTimestampDesc)
-      .reverse()
-      .map((project) => project.summary)
-      .map<CardItem>((summary) => ({
-        // TODO: Svelte v5 migration - fix type
-        component: AdoptedProposalCard as unknown as Component,
-        props: { summary },
-      }))
-  );
-
-  const cards: CardItem[] = $derived([
-    ...launchpadCards,
-    ...openProposalCards,
-    ...adoptedSnsProposalsCards,
-  ]);
+  const cards: CardItem[] = $derived([...launchpadCards, ...openProposalCards]);
 </script>
 
 <main data-tid="portfolio-page-component">
