+++ frontend/src/routes/(app)/(nns)/portfolio/+page.svelte
@@ -82,10 +82,6 @@
       swapLifecycle: SnsSwapLifecycle.Open,
       projects: $snsProjectsActivePadStore,
     })}
-    adoptedSnsProposals={filterProjectsStatus({
-      swapLifecycle: SnsSwapLifecycle.Adopted,
-      projects: $snsProjectsActivePadStore,
-    })}
     openSnsProposals={$openSnsProposalsStore}
   /></TestIdWrapper
 >
